Applied patch [ 601072 ] wxFileName::GetLongPath error & fix By Michael Fielding (mfielding)
When using wxFileName::MakeRelativeTo with a new base of "x:\", the result is incorrect. It works fine when there is a subdirectory after x:\. I tracked this to a bug in GetLongPath; here is the fix.

Also, GetLongPath will now try and get as much long path as it can, failing only when part of the path cannot be found. eg

x:\existi~1\file not existing.new becomes
x:\existing directory\filenotexisting.new

instead of just staying the same.

Also, MakeRelative no longer makes filenames lowercase when it normalises them.


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@16886 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775