Applied patch [ 549256 ] fix minor bug in widgets sample
Arnout Engelen

The EVT_TEXT_ENTER event will not be caught in the 'widgets' sample since the wxTE_PROCESS_ENTER wasn't set for the wxTextCtrls. For some reason this does work correctly in wxMSW without this flag set, which could be considered a bug or documentation bug.

This patch adds the wxTE_PROCESS_ENTER flag and thus fixes this minor bug.


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@17029 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775