Fixed bug [ 495532 ] wxFileDialog limits multiple file load
"The wxMAXPATH constant prevents the loading of
multiple file names in a multiple selection dialog
when selecting a large number of files.

The number of selected files can easily exceed the
memory allocated to store the selected file names and
paths."

I'm increasing the buffer size to 65536 which should be adequate
for most purposes. There's apparently no way to work around the
fixed buffer size. -- JACS


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_4_BRANCH@17338 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775