Don't install wxWindows until the %install step, use --inplace to facilitate this.
Change the name of the -devel package since it doesn't care about the
python version like the main pacakges do.

Instead of using directory names and wildcards in the %files section
generate an explicit file list with an %attr on each line.  This is
because the %defattr was not completly working for some reason.


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_4_BRANCH@17359 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775