Patch [ 626532 ] Fix default item/tab traversal problem
This patch fixes a problem that occurs when you have a
panel containing controls inside a dialog/frame with a
default item (i.e. default button). Currently,
<Return/Enter> is processed incorreclty as a <tab>
even if there is a default item in a parent. This patch
causes a window not to handle the VK_RETURN key if
it is not a top level window and doesn't have a default
item of it's own.

Derry Bryson


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_4_BRANCH@17610 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775