Applied FL patch from Stefan Kowski
"Enclosed there is a patch for the FL library that solves a problem
overwriting the status bar if a bar is expanded.

Cause of this is the calculation of the new window height which may be less
than 0 in some cases. In my test a value of -1 caused the window default
height to be set which was larger than the space available, therefore the
status bar was overwritten."


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@19276 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775