Applied patch [ 1178610 ] datectlg.cpp: wxDropdownButton + bug fixes Jaakko Salli
This patch modifies datectlg.cpp's wxDropdownButton in
such manner that it should be much easier to convert it
into a generic drop button class. Also, it is now more
universal in the respect that it should work on
platforms other than wxMSW and wxGTK (wxBitmapButton
being prime prerequisite).

Plus there are some bug fixes for wxDatePickerCtrlGeneric.


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@33652 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775