Bump soname in preparation for 2.6.2 release.
Could someone please rebake this with whatever the correct
version + silent hacks is.  Based on present reports I have
no idea which version and/or hacks to it are 'official', and
the crypto logic that bakefile is becoming famous for is in
shining form today:

$ bakefile_gen -fautoconf

-----------------------------------------------------------------------
This file cannot be processed with Bakefile version older than 0.1.7.
You are using Bakefile version 0.1.6. Please install the newest version
from http://bakefile.sourceforge.net.
-----------------------------------------------------------------------

It's always good to have a nice conundrum to ponder while your
code is not building...


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@34741 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775