Use AC_CHECK_TOOL rather than AC_CHECK_PROG for AR, RANLIB, STRIP, RESCOMP, and DLLTOOL.  This automatically searches for tools with a prefix of "$host_alias-" when cross compiling then falls back to the unprefixed tool name if that fails.  If the user specifies the tool using the appropriate variable (e.g. CC, CXX, AR, RANLIB, STRIP, etc.) then it will be checked as is.
Removed the major hack that did something similar to this but simply
ignored user-provided variables and used default tool names.  For example,
CC was locked to $host_alias-gcc when cross compiling even if the user
specified CC on the configure command-line.

This brings wxWidgets in line with other autoconf 2.5x projects.  The only
downside to this approach is that our old method would fail if
$host_alias-toolname was not present whereas the autoconf method will silently
use the host compiler (which is usually _not_ what the user wants).


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@35454 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775