Revert revision 1.102.2.2: date: 2002/10/13 21:03:45;  author: RR;  state: Exp;  lines: +4 -2
  Combox gets its own SetFocus()
  wxDialog() set the focus to itself when shown. This will
    the propagate to some control in the dialog. I am not
    sure if this will override preceding user's call to
    SetFocus() in one of the child windows.
=========================
It will override preceding users call to SetFocus, effectively introducing bug #1288889 (wxWindow::SetFocus does not
work on modal dialogs, and therefore wxButton::SetDefault does not work either). Reverting fixes that. Need to
investigate why it was needed in relation to Comboboxes, but SetFocus/SetDefault breakage is worser imho, hence the
revert for now.


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@35724 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775