wxX11 sockets updated to the new callback interface.
This also fixes a serious but on wxX11 sockets: the iteration of the socket event loop could crash if the socket
callback happens to remove both events, hence removing the socket from the hash list while the list is being
processed. Fixed now by lookup of the next item before the callback of the corrent one is processed.


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/branches/SOC2006_SOCKETS@40711 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775