Correct week day returned from wxCalendarCtrl::HitTest() on header click.
A combination of a wx bug in conversion from native control week days to
wxDateTime week days and a bug of native control itself when the first week
day is not Monday resulted in the day being off by one it did start with
Monday. The new code works correctly in both Monday and Sunday cases, at least
until the bug in comctl32.dll is corrected.

See comment:5 of #11057.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@61594 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775