Call wxApp::ProcessPendingEvents() from wxX11 event loop.
Now that this function is not called from idle time (because it should be
called before, see r61441), the event loop must call explicitly so do it from
wxGUIEventLoop::Dispatch() in wxX11.

A probably better (but more time-consuming) fix would be to make wxX11 use
wxEventLoopManual as there doesn't seem to be any reason not to.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@62006 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775