Remove unnecessary Mac-specific TOOLKIT_LOWERCASE setting.
TOOLKIT is never set to "MAC" at all (it is one of OSX_{CARBON,COCOA,IPHONE})
so it is not necessary to test for it. And for the other cases the special
logic for OS X ports was the same as the general case so it seems to be
unnecessary as well.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@62871 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775