Call SymCleanup() in wxStackWalker to fix problem with next call to it.
Calling SymInitialize() second and subsequent times failed (and hence so did
wxStackWalker::WalkFrom() and other methods) as we didn't call SymCleanup()
because of some problems with old versions of debughlp.dll. These problems
don't seem to exist any more so do call SymCleanup() so that we can walk the
stack more than once.

Closes #11544.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@62925 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775