Don't exclude "contrib" from wxMSW setup.exe generation.
This was presumably needed to exclude the top level contrib directory but also
seems to apply to src/tiff/contrib and excluding this directory breaks
configuration of libtiff as it looks for src/tiff/contrib/Makefile.in.

As we don't even have top level contrib any more, simply don't exclude it any
longer.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@65482 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775