No real changes, just remove an unneeded header dependency.
There was some commented out code in (core) valgen.cpp file using
wxDatePickerCtrl (from adv library). Also comment out the inclusion of
wx/datectrl.h header as it's not needed as long as this code remains commented
out and creates an unexpected dependency of a core library file on an adv
library header.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@66533 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775