Disable deprecation and other warnings in MFC sample.
VC8+ give tons of deprecation warnings for the standard functions which are
usually suppressed by wx headers but they need to be included first for the
suppression to be effective. In the MFC sample they were not resulting in many
useless warnings.

Fix this by pre-defining _CRT_SECURE_NO_WARNINGS to suppress them in the
sample itself.

Also suppress a warning about WINVER being undefined.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@66668 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775