Fix bitmap position attribute name in XRC handler.
Bitmap position attribute was called "bitmapposition" in the documentation and
in the code that generated an error message for a wrong value in it but was
inexplicably called "direction" in the code that really looked it up.

This seems to be just a straight typo from the original r61065. Surprising as
it is that it wasn't noticed before, do correct it now.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@68680 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775