Fix computation of the item offset in generic wxDataViewCtrl code.
We need to account for the expander even for the items that don't have
children when computing their offset as their parent still does have it,
contributing to the total offset.

This corrects the fixes of r68836 to also work for the leaf items.

Also rename expanderOffset variable to more clear itemOffset.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@68960 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775