Backport of r68766 (without the unit test changes).
(
Improved BMP decoding.

The BMP decoder did not handle images that are not stored upside down but straight up (in which case the height is negative). Also with RLE4 or RLE8 compressed images the 'end of scanline' RLE marker was not handled correctly. Fixed the issues and added a unit test for them.
)

Fixes #13650.


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_8_BRANCH@69732 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775