Updated out of date versions and release dates.
Mentioned more files in tn0011.txt that need version updating, and made it more clear which ones need to be changed manually. Modified inc_release to update docs/readme.txt and changed the version references in the readme.txt to 2.9.3. Also changed two references to the month of release from July to December.


git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@69920 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775