Work around impossibility to use wxCombobox in wxGrid in wxOSX.
Open the combobox when the editor is initially shown, this somehow ensures
that the combobox selection is taken into account when it is closed while
without this the selection is simply completely ignored.

This is just a hack but at least it allows to use wxGridCellChoiceEditor under
wxOSX which was previously impossible at all.

Closes #12644.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@69950 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775