Update wxNotebook selection when user changes the page in wxOSX.
wxNotebook::m_selection was only updated if the selected page was changed
programmatically but not if it was done by the user. Do update it in this case
as well, it fixes generation of the events which was broken before because of
comparison of the new selection with the incorrect current selection value in
m_selection.

git-svn-id: http://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@70340 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775