Remove unnecessary checks for OS X 10.4 and gcc 4.0 in configure.
As 10.4 is not supported any more, don't check for it. Also don't impose the
use of gcc 4.0 for compiling wxOSX/Carbon as this compiler doesn't exist any
more under OS X 10.7 but wxOSX/Carbon can be compiled using gcc 4.2 if 10.6
SDK is selected, so this change allows to build wxOSX/Carbon under 10.7.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@71950 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775